move find and getJava to parent class
